+++ server/src/main/java/com/holidayplanner/repository/JourneyRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JourneyRepository extends JpaRepository<Journey, Long> {
+public interface JourneyRepository extends JpaRepository<Journey, Integer> {
     // Find journeys by destination airport
     List<Journey> findByDestinationAirport(String destinationAirport);
 
     // Find journeys by the number of passengers
-    List<Journey> findByPassengers(int passengers);
+    List<Journey> findByPassengers(Integer passengers);
 
 }
